# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Matchbook API",
    "description": "Self-hosted agency finance ops \u2014 invoice collection, reconciliation, reporting.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/dashboard/missing-invoices": {
      "get": {
        "tags": [
          "dashboard",
          "dashboard"
        ],
        "summary": "Missing Invoice Alerts",
        "description": "Detect vendors with missing invoices based on billing cycle + last invoice date.",
        "operationId": "missing_invoice_alerts_dashboard_missing_invoices_get",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Check for missing invoices as of this period (YYYY-MM). Defaults to current month.",
              "title": "As Of"
            },
            "description": "Check for missing invoices as of this period (YYYY-MM). Defaults to current month."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissingInvoiceAlertList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/spend-summary": {
      "get": {
        "tags": [
          "dashboard",
          "dashboard"
        ],
        "summary": "Spend Summary",
        "description": "Monthly spend summary \u2014 totals, invoice count, match breakdown.",
        "operationId": "spend_summary_dashboard_spend_summary_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Period filter (YYYY-MM). Defaults to current month.",
              "title": "Period"
            },
            "description": "Period filter (YYYY-MM). Defaults to current month."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpendSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/spend-by-category": {
      "get": {
        "tags": [
          "dashboard",
          "dashboard"
        ],
        "summary": "Spend By Category",
        "description": "Expenses grouped by category for the given period.",
        "operationId": "spend_by_category_dashboard_spend_by_category_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Period filter (YYYY-MM). Defaults to current month.",
              "title": "Period"
            },
            "description": "Period filter (YYYY-MM). Defaults to current month."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategorySpendList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/spend-by-vendor": {
      "get": {
        "tags": [
          "dashboard",
          "dashboard"
        ],
        "summary": "Spend By Vendor",
        "description": "Expenses grouped by vendor for the given period.",
        "operationId": "spend_by_vendor_dashboard_spend_by_vendor_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Period filter (YYYY-MM). Defaults to current month.",
              "title": "Period"
            },
            "description": "Period filter (YYYY-MM). Defaults to current month."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorSpendList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/vat-summary": {
      "get": {
        "tags": [
          "dashboard",
          "dashboard"
        ],
        "summary": "Vat Summary",
        "description": "VAT breakdown by rate for the given period.",
        "operationId": "vat_summary_dashboard_vat_summary_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Period filter (YYYY-MM). Defaults to current month.",
              "title": "Period"
            },
            "description": "Period filter (YYYY-MM). Defaults to current month."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VATSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/mom-comparison": {
      "get": {
        "tags": [
          "dashboard",
          "dashboard"
        ],
        "summary": "Mom Comparison",
        "description": "Month-over-month spend comparison for the given year.",
        "operationId": "mom_comparison_dashboard_mom_comparison_get",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Year to compare months for. Defaults to current year.",
              "title": "Year"
            },
            "description": "Year to compare months for. Defaults to current year."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoMComparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/tax-summary": {
      "get": {
        "tags": [
          "dashboard",
          "dashboard"
        ],
        "summary": "Tax Summary",
        "description": "Tax and non-invoiceable transactions for the given period (bank fees, tax payments, etc.).",
        "operationId": "tax_summary_dashboard_tax_summary_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Period filter (YYYY-MM). Defaults to current month.",
              "title": "Period"
            },
            "description": "Period filter (YYYY-MM). Defaults to current month."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/earnings": {
      "get": {
        "tags": [
          "dashboard",
          "dashboard"
        ],
        "summary": "Earnings",
        "description": "Incoming payments (credits) for the given period.",
        "operationId": "earnings_dashboard_earnings_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Period filter (YYYY-MM). Defaults to current month.",
              "title": "Period"
            },
            "description": "Period filter (YYYY-MM). Defaults to current month."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarningsSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/withholdings": {
      "get": {
        "tags": [
          "dashboard",
          "dashboard"
        ],
        "summary": "Withholdings",
        "description": "Owner withdrawals against earnings for the given period.",
        "operationId": "withholdings_dashboard_withholdings_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Period filter (YYYY-MM). Defaults to current month.",
              "title": "Period"
            },
            "description": "Period filter (YYYY-MM). Defaults to current month."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithholdingSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BillingCycle": {
        "type": "string",
        "enum": [
          "monthly",
          "annual",
          "one_off"
        ],
        "title": "BillingCycle"
      },
...
      "CategorySpendList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CategorySpend"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CategorySpendList"
      },
...
      "EarningsSummary": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period"
          },
          "total_earnings": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Earnings"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EarningTransaction"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "period",
          "total_earnings",
          "transaction_count",
          "items"
        ],
        "title": "EarningsSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MissingInvoiceAlert": {
        "properties": {
          "vendor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Vendor Id"
          },
          "vendor_name": {
            "type": "string",
            "title": "Vendor Name"
          },
          "billing_cycle": {
            "$ref": "#/components/schemas/BillingCycle"
          },
          "last_invoice_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Invoice Period"
          },
          "expected_period": {
            "type": "string",
            "title": "Expected Period"
          }
        },
        "type": "object",
        "required": [
          "vendor_id",
          "vendor_name",
          "billing_cycle",
          "expected_period"
        ],
        "title": "MissingInvoiceAlert"
      },
      "MissingInvoiceAlertList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MissingInvoiceAlert"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "MissingInvoiceAlertList"
      },
      "MoMComparison": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MonthlySpend"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "MoMComparison"
      },
...
      "SpendSummary": {
        "properties": {
          "total_spend_excl": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Spend Excl"
          },
          "total_vat": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Vat"
          },
          "total_spend_incl": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Spend Incl"
          },
          "invoice_count": {
            "type": "integer",
            "title": "Invoice Count"
          },
          "matched_count": {
            "type": "integer",
            "title": "Matched Count"
          },
          "unmatched_count": {
            "type": "integer",
            "title": "Unmatched Count"
          }
        },
        "type": "object",
        "required": [
          "total_spend_excl",
          "total_vat",
          "total_spend_incl",
          "invoice_count",
          "matched_count",
          "unmatched_count"
        ],
        "title": "SpendSummary"
      },
...
      "TaxSummary": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period"
          },
          "total_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count"
          },
          "by_category": {
            "items": {
              "$ref": "#/components/schemas/TaxCategoryBreakdown"
            },
            "type": "array",
            "title": "By Category"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaxTransaction"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "period",
          "total_amount",
          "transaction_count",
          "by_category",
          "items"
        ],
        "title": "TaxSummary"
      },
...
      "VATSummary": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VATBreakdown"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "VATSummary"
      },
...
      "VendorSpendList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VendorSpend"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "VendorSpendList"
      },
...
      "WithholdingSummary": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period"
          },
          "total_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WithholdingTransaction"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "period",
          "total_amount",
          "transaction_count",
          "items"
        ],
        "title": "WithholdingSummary"
      },
...
    }
  }
}